refactor: split up Round component
smaller components are easier to maintain

ref https://github.com/mlevison/high-performance-teams-game/issues/65
