+++ src/components/FinalResults.tsx
