+++ src/components/Round/index.tsx
@@ -1,0 +1,2 @@
+export { default } from './Round';
+export { default as Results } from './Results';
