+++ src/components/Round/types.ts
@@ -1,0 +1,1 @@
+export type View = 'welcome' | 'actions' | 'results';
