+++ src/components/index.ts
@@ -1,5 +1,5 @@
-export { default as Results } from './Results';
-export { default as Round } from './Round';
+export { default as FinalResults } from './FinalResults';
+export { default as Round, Results } from './Round';
 export { default as Actions } from './Actions';
 export { default as Status } from './Status';
 export { default as Header } from './Header';
